BE/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.10.0
python-dotenv==1.0.0
supabase==2.0.2
pinecone
google-generativeai==0.3.2
python-multipart==0.0.6
httpx>=0.24.0,<0.25.0
aiohttp==3.9.1
PyPDF2==3.0.1
pypdf==3.17.1
pdfplumber==0.10.3
langchain-text-splitters==0.0.1
reportlab==4.0.4
